refactor: Replace Task in Lesson with task ids and remove RawLesson
This allows us lazily fetch tasks when they are actually needed
