+++ lib/data/repositories/lessons/lessons_repository.dart
@@ -9,19 +9,6 @@ class LessonsRepository {
   final RestClient restClient;
 
   Future<List<Lesson>> getLessons() async {
-    final lessons = await restClient.getLessons();
-    final tasks = await restClient.getTasks();
-    final tasksMap = {for (final task in tasks) task.id: task};
-
-    return lessons
-        .map(
-          (lesson) => Lesson(
-            id: lesson.id,
-            title: lesson.title,
-            pages: lesson.pages,
-            tasks: lesson.taskIds.map((id) => tasksMap[id]!).toList(),
-          ),
-        )
-        .toList();
+    return restClient.getLessons();
   }
 }
