+++ lib/data/services/rest_client.dart
@@ -1,5 +1,5 @@
 import 'package:dio/dio.dart';
-import 'package:lessons_tasks_assignment/data/repositories/lessons/models/raw_lesson.dart';
+import 'package:lessons_tasks_assignment/domain/lesson.dart';
 import 'package:lessons_tasks_assignment/domain/task.dart';
 import 'package:retrofit/retrofit.dart';
 
@@ -15,7 +15,7 @@ abstract class RestClient {
       'https://raw.githubusercontent.com/Peetee06/lessons-tasks-assignment/refs/heads/main/';
 
   @GET('/mock_lessons.json')
-  Future<List<RawLesson>> getLessons();
+  Future<List<Lesson>> getLessons();
 
   @GET('/mock_tasks.json')
   Future<List<Task>> getTasks();
