+++ lib/domain/lesson.dart
@@ -1,9 +1,9 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:lessons_tasks_assignment/common/typedefs.dart';
 import 'package:lessons_tasks_assignment/domain/page.dart';
-import 'package:lessons_tasks_assignment/domain/task.dart';
 
 part 'lesson.freezed.dart';
+part 'lesson.g.dart';
 
 @freezed
 class Lesson with _$Lesson {
@@ -11,6 +11,8 @@ class Lesson with _$Lesson {
     required String id,
     required Localizations title,
     required List<Page> pages,
-    required List<Task> tasks,
+    required List<String> taskIds,
   }) = _Lesson;
+
+  factory Lesson.fromJson(Map<String, dynamic> json) => _$LessonFromJson(json);
 }
