+++ test/features/lesson/view/lesson_view_test.dart
@@ -4,7 +4,6 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:lessons_tasks_assignment/domain/content_component.dart';
 import 'package:lessons_tasks_assignment/domain/lesson.dart';
 import 'package:lessons_tasks_assignment/domain/page.dart';
-import 'package:lessons_tasks_assignment/domain/task.dart';
 import 'package:lessons_tasks_assignment/features/lesson/cubit/lesson_cubit.dart';
 import 'package:lessons_tasks_assignment/features/lesson/cubit/lesson_state.dart';
 import 'package:lessons_tasks_assignment/features/lesson/view/lesson_view.dart';
@@ -72,7 +71,7 @@ void main() {
             ],
           ),
         ],
-        tasks: [],
+        taskIds: [],
       );
       when(cubit.state).thenReturn(const LessonLoaded(lesson));
       await pumpTestWidget(tester);
@@ -87,7 +86,7 @@ void main() {
         id: '1',
         title: {'en': 'Test Lesson', 'de': 'Test Lektion'},
         pages: [],
-        tasks: [],
+        taskIds: [],
       );
       when(cubit.state).thenReturn(const LessonLoaded(lesson));
       await pumpTestWidget(tester);
@@ -102,14 +101,7 @@ void main() {
         id: '1',
         title: {'en': 'Test Lesson', 'de': 'Test Lektion'},
         pages: [],
-        tasks: [
-          Task(
-            id: '1',
-            question: {'en': 'Test Question', 'de': 'Test Frage'},
-            options: [],
-            correctAnswerIds: [],
-          ),
-        ],
+        taskIds: ['1'],
       );
       when(cubit.state).thenReturn(const LessonLoaded(lesson));
       await pumpTestWidget(tester);
